+++ ios/volume_controller/Sources/volume_controller/VolumeControllerPlugin.swift
@@ -3,9 +3,9 @@ import Flutter
 import UIKit
 
 public class VolumeControllerPlugin: NSObject, FlutterPlugin {
-  private static let audioSession = AVAudioSession.sharedInstance()
-  private static let volumeController = VolumeController(audioSession: audioSession)
-  private static let volumeListener = VolumeListener(audioSession: audioSession)
+  private let audioSession = AVAudioSession.sharedInstance()
+  private lazy var volumeController = VolumeController(audioSession: audioSession)
+  private lazy var volumeListener = VolumeListener(audioSession: audioSession)
 
   private func invalidArgumentsResult() -> FlutterError {
     FlutterError(
@@ -15,21 +15,17 @@ public class VolumeControllerPlugin: NSObject, FlutterPlugin {
   }
 
   public static func register(with registrar: FlutterPluginRegistrar) {
-    // Method Channel
     let methodChannel = FlutterMethodChannel(
       name: ChannelName.methodChannel, binaryMessenger: registrar.messenger())
     let instance = VolumeControllerPlugin()
     registrar.addMethodCallDelegate(instance, channel: methodChannel)
 
-    // Volume Listener Event Channel
-    let eventChannel: FlutterEventChannel = FlutterEventChannel(
+    let eventChannel = FlutterEventChannel(
       name: ChannelName.eventChannel,
       binaryMessenger: registrar.messenger())
-    eventChannel.setStreamHandler(volumeListener)
+    eventChannel.setStreamHandler(instance.volumeListener)
 
-    // Application Life Cycle
     registrar.addApplicationDelegate(instance)
-    // Scene Life Cycle (iOS 13 and above)
     if #available(iOS 13.0, *) {
       registrar.addSceneDelegate(instance)
     }
@@ -38,8 +34,7 @@ public class VolumeControllerPlugin: NSObject, FlutterPlugin {
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
     switch call.method {
     case MethodName.getVolume:
-      let volume = VolumeControllerPlugin.volumeController.getVolume()
-      result(volume)
+      result(volumeController.getVolume())
     case MethodName.setVolume:
       let arg = call.arguments as? [String: Any]
       let volume = arg?[MethodArgument.volume] as? Double
@@ -50,12 +45,11 @@ public class VolumeControllerPlugin: NSObject, FlutterPlugin {
         return
       }
 
-      VolumeControllerPlugin.volumeController.setVolume(
-        volume: Float(volume), showSystemUI: showSystemUI)
-      result(nil)
+      volumeController.setVolume(volume: Float(volume), showSystemUI: showSystemUI) { error in
+        result(error)
+      }
     case MethodName.isMuted:
-      let isMuted = VolumeControllerPlugin.volumeController.isMuted()
-      result(isMuted)
+      result(volumeController.isMuted())
     case MethodName.setMute:
       let arg = call.arguments as? [String: Any]
       let isMute = arg?[MethodArgument.isMute] as? Bool
@@ -66,38 +60,28 @@ public class VolumeControllerPlugin: NSObject, FlutterPlugin {
         return
       }
 
-      VolumeControllerPlugin.volumeController.setMute(isMute: isMute, showSystemUI: showSystemUI)
-      result(nil)
+      volumeController.setMute(isMute: isMute, showSystemUI: showSystemUI) { error in
+        result(error)
+      }
     default:
       result(FlutterMethodNotImplemented)
     }
-  }
-
-  private func resumeVolumeListener() {
-    guard VolumeControllerPlugin.volumeListener.isObservingVolume else {
-      return
-    }
-
-    VolumeControllerPlugin.audioSession.setAudioSessionCategory()
-    VolumeControllerPlugin.audioSession.activateAudioSession()
-    VolumeControllerPlugin.volumeListener.sendVolumeChangeEvent()
   }
 }
 
 extension VolumeControllerPlugin: FlutterApplicationLifeCycleDelegate {
   public func applicationWillEnterForeground(_ application: UIApplication) {
-    // On iOS 13 and above, the scene delegate's sceneWillEnterForeground will be called instead of this method.
     if #available(iOS 13.0, *) {
       return
     }
 
-    resumeVolumeListener()
+    volumeListener.resumeVolumeObservation()
   }
 }
 
 @available(iOS 13.0, *)
 extension VolumeControllerPlugin: FlutterSceneLifeCycleDelegate {
   public func sceneWillEnterForeground(_ scene: UIScene) {
-    resumeVolumeListener()
+    volumeListener.resumeVolumeObservation()
   }
 }