+++ windows/include/volume_controller/volume_controller.h
@@ -4,10 +4,14 @@
 #include <windows.h>
 #include <mmdeviceapi.h>
 #include <endpointvolume.h>
-#include <iostream>
+#include <mutex>
+
+#include "volume_callback.h"
 
 namespace volume_controller
 {
+    class DeviceNotificationClient;
+
     class VolumeController
     {
     public:
@@ -25,6 +29,10 @@ namespace volume_controller
 
         bool SetMute(bool isMute);
 
+        bool RegisterVolumeNotification(volume_callback::VolumeCallback *callback);
+
+        void DisposeVolumeNotification();
+
     private:
         VolumeController() = default;
         ~VolumeController() = default;
@@ -32,8 +40,22 @@ namespace volume_controller
         VolumeController(const VolumeController &) = delete;
         VolumeController &operator=(const VolumeController &) = delete;
 
+        bool InitializeLocked();
+        void DisposeLocked();
+        bool BindDefaultEndpointLocked();
+        void ReleaseEndpointLocked();
+        void OnDefaultDeviceChanged();
+
+        friend class DeviceNotificationClient;
+
+        std::recursive_mutex mutex_;
+        int init_count_ = 0;
+        bool com_needs_uninit_ = false;
+        IMMDeviceEnumerator *pEnumerator_ = nullptr;
         IAudioEndpointVolume *pVolume_ = nullptr;
+        volume_callback::VolumeCallback *pCallback_ = nullptr;
+        DeviceNotificationClient *pDeviceClient_ = nullptr;
     };
 }
 
-#endif // VOLUME_CONTROLLER_H_
+#endif